airapp1/requirements.txt
Django==1.10.5
django_import_export==0.5.1
djangorestframework==3.5.4
